Precompiled headers

+++ src/Intelligence/EStrat.h
@@ -6,7 +6,7 @@ class EStrat : public CSingleton<EStrat>
 	friend class CSingleton<EStrat>;
 
 public :
-	int probability_of_rush();
-	int probability_of_proxy();
-	int probability_of_canonrush();
+	double probability_of_rush();
+	double probability_of_proxy();
+	double probability_of_canonrush();
 };
